+++ doc/CHANGELOG.md
@@ -7,6 +7,12 @@
         of the response parameters. This is commonly termed as rpHash.
       - Added option **-S**, **\--session** to specify to specify an auxiliary
         session for auditing and or encryption/decryption of the parameters.
+      - Added option **-n**, **\--name** to specify the name of the nvindex in
+        hex bytes. This is used when cpHash is calculated without dispatching
+        the TPM2_NV_Write command to the TPM.
+      - Added option **\-signer-name** to specify the signer name in hex bytes.
+        This is used when cpHash is calculated without dispatching the
+        TPM2_NV_Write command to the TPM.
 
   * tpm2_nvwrite:
       - Added option **\--rphash**=_FILE_ to specify ile path to record the hash